♻️ 🐛  Refactor quality evaluation handling by moving getColorForEvaluation and qualityList to a new quality-list module, and update getScheduledTunesOverview to return ITuneOverviewScheduled type, which properly contains the recall_eval field, fixes bug per dissapearing evals.
